requirements.txt
pandas>=1.5.0,<1.6.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.2.0,<1.4.0
nltk>=3.8.0
textblob>=0.17.0
streamlit>=1.22.0
requests>=2.32.3
beautifulsoup4>=4.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
smolagents>=0.4.0
python-dotenv>=1.0.0
openai>=1.0.0
transformers>=4.30.0
spacy>=3.7.0
vaderSentiment>=3.3.2
torch>=2.0.0
emoji>=2.8.0
pytz>=2023.3
google-api-python-client>=2.100.0
python-dateutil>=2.8.2

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# Additional ML and data science
joblib>=1.2.0
scipy>=1.10.0

# Database support
# sqlite3 is built into Python

# Caching
redis>=4.5.0
diskcache>=5.6.0

# Monitoring and logging
loguru>=0.7.0
psutil>=5.9.0

# API and web framework enhancements
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0

# Testing data generation
faker>=18.0.0
